+++ src/index.ts
@@ -1,23 +1,11 @@
 import express from "express";
 import cors from "cors";
 import path from "path";
-import { expressMiddlewares } from "./routes";
+import { middlewares } from "./routes";
+import { corsOptions } from "./corsConfig";
 
 const app = express();
 const port = process.env.APP_PORT || 5005;
-
-const allowedOrigins = [
-  process.env.FRONTEND_URL || "http://localhost:3000",
-  process.env.TGC_RELEASE_URL || "http://localhost:3000",
-  process.env.TGC_PROD_URL || "http://localhost:3000",
-];
-const corsOptions = {
-  origin: allowedOrigins,
-  methods: "GET,HEAD,PUT,PATCH,POST,DELETE,OPTIONS",
-  allowedHeaders: ["Content-Type", "Authorization"],
-  credentials: true,
-  optionsSuccessStatus: 200,
-};
 
 app.use(cors(corsOptions));
 app.use(express.json({ limit: "50mb" }));
@@ -30,7 +18,7 @@ app.use(
 app.get("/", (req, res) => {
   res.send("Bienvenue sur Mega-S3!");
 });
-expressMiddlewares(app);
+middlewares(app);
 app.listen(port, () => {
   console.log(`🚀 Server ready at port ${port} 🚀`);
 });
